requirements.txt
alembic==1.16.4
asarPy==1.0.1
blinker==1.9.0
Brotli==1.1.0
certifi==2025.7.9
cffi==1.17.1
charset-normalizer==3.4.2
click==8.1.7
cloudinary==1.44.1
colorama==0.4.6
contourpy==1.3.3
cssselect2==0.8.0
cycler==0.12.1
Flask==3.1.0
Flask-Cors==5.0.0
Flask-JWT-Extended==4.6.0
flask-marshmallow==1.2.1
Flask-Migrate==4.1.0
Flask-SQLAlchemy==3.1.1
fonttools==4.59.0
greenlet==3.1.1
gunicorn==23.0.0
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.4
kiwisolver==1.4.8
Mako==1.3.10
MarkupSafe==3.0.2
marshmallow==3.23.1
marshmallow-sqlalchemy==1.1.0
matplotlib==3.10.3
numpy==2.3.2
packaging==24.2
pillow==11.3.0
psycopg2-binary==2.9.10
pycparser==2.22
pydyf==0.11.0
PyJWT==2.10.1
pyparsing==3.2.3
pyphen==0.17.2
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
requests==2.32.4
six==1.17.0
SQLAlchemy==2.0.36
tinycss2==1.4.0
tinyhtml5==2.0.0
typing_extensions==4.12.2
tzdata==2025.2
urllib3==2.5.0
waitress==3.0.2
weasyprint==66.0
webencodings==0.5.1
Werkzeug==3.1.3
zopfli==0.2.3.post1
